requirements.txt
# =========================================================
# PYTHON CONSTRAINT
# =========================================================
# python_version < "3.11"

# =========================================================
# CORE APP
# =========================================================
streamlit==1.52.2
pillow==12.0.0
numpy==1.26.4
opencv-python==4.8.1.78
pandas==2.3.3

# =========================================================
# TENSORFLOW (SINGLE RUNTIME — REQUIRED)
# =========================================================
tensorflow==2.15.1
ml-dtypes==0.3.1
h5py==3.10.0
protobuf==4.25.3
tensorboard==2.15.2

# =========================================================
# TORCH (CPU)
# =========================================================
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# =========================================================
# ML / SCIENCE
# =========================================================
scikit-learn==1.7.2
scipy==1.15.3
matplotlib==3.10.8
seaborn==0.13.2
scikit-image==0.25.2
shap==0.49.1
lime==0.2.0.1

# =========================================================
# NLP / LLM
# =========================================================
transformers==4.57.3
sentence-transformers==5.2.0
tokenizers==0.22.2
tiktoken==0.12.0
huggingface-hub==0.36.0

# =========================================================
# LANGCHAIN
# =========================================================
langchain==1.2.3
langchain-community==0.4.1
langchain-core==1.2.6
langchain-text-splitters==1.1.0
langchain-huggingface==1.2.0
langchain-google-genai==4.1.3
langgraph==1.0.5
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.2

# =========================================================
# FASTAPI (OPTIONAL)
# =========================================================
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
python-multipart==0.0.21

# =========================================================
# UTILS
# =========================================================
requests==2.32.5
rich==14.3.1
tqdm==4.67.1
PyYAML==6.0.3
python-dotenv==1.2.1
